requirements.txt
crewai>=0.28.8
langchain-google-genai>=1.0.3
streamlit>=1.33.0
python-dotenv>=1.0.1
